+++ src/commands/validate-tasks.ts
@@ -3,6 +3,7 @@ import { BaseCliCommand } from './base-command';
 import { IPerformTasksCommandArgs } from './perform-tasks';
 import { BuildConfiguration } from '~build-tasks/build-configuration';
 import { BuildRunner } from '~build-tasks/build-runner';
+import { Validator } from '~parser/validator';
 
 const commandName = 'validate-tasks <templateFile>';
 const commandDescription = 'Will validate the tasks file, including configured tasks';
@@ -22,7 +23,7 @@ export class ValidateTasksCommand extends BaseCliCommand<IPerformTasksCommandArg
         command.option('--organization-file [organization-file]', 'organization file used for organization bindings');
         command.option('--max-concurrent-tasks <max-concurrent-tasks>', 'maximum number of tasks to be executed concurrently', 1);
         command.option('--max-concurrent-stacks <max-concurrent-stacks>', 'maximum number of stacks (within a task) to be executed concurrently', 1);
-        command.option('--failed-tasks-tolerance <failed-tasks-tolerance>', 'the number of failed tasks after which execution stops', 0);
+        command.option('--failed-tasks-tolerance <failed-tasks-tolerance>', 'the number of failed tasks after which execution stops', 99);
         command.option('--failed-stacks-tolerance <failed-stacks-tolerance>', 'the number of failed stacks (within a task) after which execution stops', 0);
         command.option('--organization-file [organization-file]', 'organization file used for organization bindings');
         command.option('--parameters [parameters]', 'parameters used when creating build tasks from tasks file');
@@ -33,10 +34,16 @@ export class ValidateTasksCommand extends BaseCliCommand<IPerformTasksCommandArg
     public async performCommand(command: IPerformTasksCommandArgs): Promise<void> {
         const tasksFile = command.tasksFile;
 
+
+        Validator.validatePositiveInteger(command.maxConcurrentStacks, 'maxConcurrentStacks');
+        Validator.validatePositiveInteger(command.failedStacksTolerance, 'failedStacksTolerance');
+        Validator.validatePositiveInteger(command.maxConcurrentTasks, 'maxConcurrentTasks');
+        Validator.validatePositiveInteger(command.failedTasksTolerance, 'failedTasksTolerance');
+
         command.parsedParameters = this.parseCfnParameters(command.parameters);
         const config = new BuildConfiguration(tasksFile, command.parsedParameters);
 
         const validationTasks = config.enumValidationTasks(command);
-        await BuildRunner.RunValidationTasks(validationTasks, command.verbose === true ,1 , 999);
+        await BuildRunner.RunValidationTasks(validationTasks, command.verbose === true , command.maxConcurrentTasks, command.failedTasksTolerance);
     }
 }