+++ test/integration-tests/scenario-update-stacks-with-param-expressions.test.ts
@@ -81,7 +81,6 @@ describe('when importing value from another stack', () => {
         expect(parameter.ParameterValue).toBe(output.OutputValue);
     })
 
-
     test('Stack parameter with !Ref gets resolved ', () =>{
         expect(describeBucketRoleStack).toBeDefined();
 
@@ -147,6 +146,75 @@ describe('when importing value from another stack', () => {
         expect(parameter.ParameterValue).toBe(output.OutputValue + '-postfix');
     })
 
+    test('FindInMap gets resolved properly', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'findInMap1');
+        expect(parameter.ParameterValue).toBe('MyVal1');
+    })
+
+    test('FindInMap with parameter gets resolved properly', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'findInMap1');
+        expect(parameter.ParameterValue).toBe('MyVal1');
+    })
+
+    test('md5 gets resolved properly', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'md5');
+        expect(parameter.ParameterValue).toBe('5377a3405f914e569220a5a65e318f9a');
+    })
+
+    test('readFile gets resolved properly', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'readFile');
+        expect(parameter.ParameterValue).toBe('contents of file');
+    })
+
+    test('md5 and readFile gets resolved properly', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'md5readFile');
+        expect(parameter.ParameterValue).toBe('93b885adfe0da089cdf634904fd59f71');
+    })
+
+    test('select gets resolved properly', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'select');
+        expect(parameter.ParameterValue).toBe('three');
+    })
+
+    test('select combined with find-in-map gets resolved properly', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'selectFindInMap');
+        expect(parameter.ParameterValue).toBe('MyVal1');
+    })
+
+    test('ref to OrganizationRoot resolves to physical id of root', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'refToRoot');
+        expect(parameter.ParameterValue).toBe('r-kvte');
+    })
+
+    test('JsonString resolves to minimized json string', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'jsonString1');
+        expect(parameter.ParameterValue).toBe('{"att":1,"otherAtt":"2"}');
+    })
+
+    test('JsonString can be used with ReadFile', () =>{
+        expect(describeBucketRoleStack).toBeDefined();
+
+        const parameter = describeBucketRoleStack.Stacks[0].Parameters.find(x=>x.ParameterKey === 'jsonString2');
+        expect(parameter.ParameterValue).toBe('{"key":"val"}');
+    })
     test('cleanup removes deployed stacks', () => {
         expect(stacksAfterCleanup.StackSummaries.find(x=>x.StackName === 'my-scenario-export-bucket')).toBeUndefined();
         expect(stacksAfterCleanup.StackSummaries.find(x=>x.StackName === 'my-scenario-export-bucket-role')).toBeUndefined();