+++ test/unit-tests/writer/default-template-writer.test.ts
@@ -13,7 +13,7 @@ describe('when writing template for organization', () => {
         jest.spyOn(organization, 'initialize').mockImplementation();
         organization.organization = {MasterAccountId: '111111111111'};
         organization.roots = [{ Id: 'o-root', Policies: [], OrganizationalUnits: [] }];
-        organization.masterAccount = {Id: '111111111111',  Name: 'Organization Master Account', ParentId: 'o-root', Email: 'email@someplace.com', Policies: [], Type: 'Account'};
+        organization.masterAccount = {Id: '111111111111',  Name: 'Organization Master Account', ParentId: 'o-root', Email: 'email@someplace.com', Policies: [], Type: 'Account', Tags: {key : 'val'}};
         organization.organizationalUnits = [];
         organization.accounts = [];
         organization.policies = [];
@@ -64,6 +64,13 @@ describe('when writing template for organization', () => {
             const defaultTemplate = await templateWriter.generateDefaultTemplate();
             const root = TemplateRoot.createFromContents(defaultTemplate.template);
             expect(root.organizationSection.masterAccount).toBeDefined();
+        });
+
+        test('generated template contains master account tags', async () => {
+            const defaultTemplate = await templateWriter.generateDefaultTemplate();
+            const root = TemplateRoot.createFromContents(defaultTemplate.template);
+            expect(root.organizationSection.masterAccount).toBeDefined();
+            expect(root.organizationSection.masterAccount.tags).toBeDefined();
         });
 
         test('generated template contains master account email', async () => {